% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize-strategy.R
\name{binarize_largest_gap}
\alias{binarize_largest_gap}
\title{Largest-gap binarization strategy.}
\usage{
binarize_largest_gap()
}
\value{
A \code{binarize_strategy} object.
}
\description{
Creates a binarization strategy that reduces multiclass nodes to binary
by projecting group means and splitting at the largest gap. This is the
default binarization strategy.
}
\examples{
binarize_largest_gap()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}
}
