% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint-strategy.R
\name{cutpoint_mean_of_means}
\alias{cutpoint_mean_of_means}
\title{Mean-of-means split cutpoint strategy.}
\usage{
cutpoint_mean_of_means()
}
\value{
A \code{cutpoint_strategy} object.
}
\description{
Creates a split cutpoint strategy that splits at the midpoint between group
means. This is the default (and currently only) split cutpoint.
}
\examples{
cutpoint_mean_of_means()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}, \code{\link{pp_pda}}, \code{\link{vars_uniform}}
}
