% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{dr_noop}
\alias{dr_noop}
\title{No-op dimensionality reduction strategy.}
\usage{
dr_noop()
}
\value{
A \code{dr_strategy} object.
}
\description{
Creates a dimensionality reduction strategy that uses all variables at each
split. This is the default for single trees (\code{\link{pptr}}).
}
\examples{
dr_noop()

}
\seealso{
\code{\link{pptr}}, \code{\link{dr_uniform}}, \code{\link{pp_pda}}, \code{\link{sr_mean_of_means}}
}
