% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{dr_uniform}
\alias{dr_uniform}
\title{Uniform random variable selection strategy.}
\usage{
dr_uniform(n_vars = NULL, p_vars = NULL)
}
\arguments{
\item{n_vars}{The number of variables to consider at each split (integer).
Cannot be used together with \code{p_vars}.}

\item{p_vars}{The proportion of variables to consider at each split
(number between 0 and 1, exclusive). Resolved to an integer count when
the number of features is known. Cannot be used together with \code{n_vars}.}
}
\value{
A \code{dr_strategy} object.
}
\description{
Creates a dimensionality reduction strategy that randomly selects a subset
of variables at each split. Used with \code{\link{pprf}} for random forests.
}
\examples{
# Select 2 variables at each split
dr_uniform(n_vars = 2)

# Select half the variables at each split
dr_uniform(p_vars = 0.5)

# Use with pprf
pprf(Type ~ ., data = iris, dr = dr_uniform(n_vars = 2))

}
\seealso{
\code{\link{pprf}}, \code{\link{dr_noop}}, \code{\link{pp_pda}}, \code{\link{sr_mean_of_means}}
}
