% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPForest.R
\name{formula.PPForest}
\alias{formula.PPForest}
\title{Extracts the formula used to train a PPForest model.}
\usage{
\method{formula}{PPForest}(x, ...)
}
\arguments{
\item{x}{A PPForest model.}

\item{...}{(unused) other parameters typically passed to formula.}
}
\value{
The formula used to train the model.
}
\description{
Extracts the formula used to train a PPForest model.
}
\examples{
model <- PPForest(Type ~ ., data = iris)
formula(model)
}
\seealso{
\code{\link{PPForest}} for training, \code{\link{predict.PPForest}}, \code{\link{summary.PPForest}}
}
