% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmodel.R
\name{formula.ppmodel}
\alias{formula.ppmodel}
\title{Formula extractor for ppforest2 models.}
\usage{
\method{formula}{ppmodel}(x, ...)
}
\arguments{
\item{x}{A \code{pptr} or \code{pprf} model.}

\item{...}{Unused.}
}
\value{
The formula the model was trained with, or `NULL` for matrix-interface fits.
}
\description{
Formula extractor for ppforest2 models.
}
