% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprf.R
\name{formula.pprf}
\alias{formula.pprf}
\title{Extracts the formula used to train a pprf model.}
\usage{
\method{formula}{pprf}(x, ...)
}
\arguments{
\item{x}{A pprf model.}

\item{...}{(unused) other parameters typically passed to formula.}
}
\value{
The formula used to train the model.
}
\description{
Extracts the formula used to train a pprf model.
}
\examples{
model <- pprf(Type ~ ., data = iris)
formula(model)
}
\seealso{
\code{\link{pprf}} for training, \code{\link{predict.pprf}}, \code{\link{summary.pprf}}
}
