% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptr.R
\name{formula.pptr}
\alias{formula.pptr}
\title{Extracts the formula used to train a pptr model.}
\usage{
\method{formula}{pptr}(x, ...)
}
\arguments{
\item{x}{A pptr model.}

\item{...}{(unused) other parameters typically passed to formula.}
}
\value{
The formula used to train the model.
}
\description{
Extracts the formula used to train a pptr model.
}
\examples{
model <- pptr(Type ~ ., data = iris)
formula(model)
}
\seealso{
\code{\link{pptr}} for training, \code{\link{predict.pptr}}, \code{\link{summary.pptr}}
}
