% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-strategy.R
\name{grouping_by_label}
\alias{grouping_by_label}
\title{Label-based grouping strategy.}
\usage{
grouping_by_label()
}
\value{
A \code{grouping_strategy} object.
}
\description{
Creates a grouping strategy that routes all observations of a group to the
same child node. This is the default grouping strategy for classification.
}
\examples{
grouping_by_label()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}
}
