% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{Iris Flower Dataset}
\format{
A data frame with 150 rows and 5 variables:
\describe{
  \item{Sepal.Length}{Length of the sepal in cm.}
  \item{Sepal.Width}{Width of the sepal in cm.}
  \item{Petal.Length}{Length of the petal in cm.}
  \item{Petal.Width}{Width of the petal in cm.}
  \item{Type}{A factor with levels `setosa`, `versicolor`, and `virginica` indicating the species of each observation.}
}
}
\source{
Fisher, R.A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179-188.
}
\description{
The Iris flower dataset or Fisher's Iris dataset is a multivariate dataset introduced by Sir Ronald Fisher in 1936 as an example of discriminant analysis. It includes measurements for 150 iris flowers from three species, with 50 from each species. The measurements include the length and the width of the sepals and petals.
}
