% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf-strategy.R
\name{leaf_majority_vote}
\alias{leaf_majority_vote}
\title{Majority-vote leaf strategy.}
\usage{
leaf_majority_vote()
}
\value{
A \code{leaf_strategy} object.
}
\description{
Creates a leaf strategy that assigns the majority group label as the
leaf prediction. When groups are tied, the smallest label wins.
This is the default leaf strategy for classification.
}
\examples{
leaf_majority_vote()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}
}
