% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf-strategy.R
\name{leaf_mean_response}
\alias{leaf_mean_response}
\title{Mean-response leaf strategy.}
\usage{
leaf_mean_response()
}
\value{
A \code{leaf_strategy} object.
}
\description{
Creates a leaf strategy that predicts the mean of the continuous response
values for the observations in the leaf. Used for regression trees; requires
a numeric response.
}
\examples{
leaf_mean_response()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}, \code{\link{leaf_majority_vote}}
}
