% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmodel.R
\name{oob_samples}
\alias{oob_samples}
\title{Out-of-bag row indices per tree.}
\usage{
oob_samples(model)
}
\arguments{
\item{model}{A \code{pprf} forest model.}
}
\value{
A list of integer vectors, one per tree.
}
\description{
Returns a list where element `i` is the integer vector of row indices
(1-based) that were **not** in the bootstrap sample of tree `i`.
}
\seealso{
\code{\link{bag_samples}}, \code{\link{oob_error}}
}
