% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-strategy.R
\name{partition_by_group}
\alias{partition_by_group}
\title{Group-based partition strategy.}
\usage{
partition_by_group()
}
\value{
A \code{partition_strategy} object.
}
\description{
Creates a partition strategy that routes all observations of a group to the
same child node. This is the default partition strategy.
}
\examples{
partition_by_group()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}
}
