% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp-strategy.R
\name{pp_pda}
\alias{pp_pda}
\title{PDA projection pursuit strategy.}
\usage{
pp_pda(lambda = 0)
}
\arguments{
\item{lambda}{A regularization parameter between 0 and 1.
If \code{lambda = 0}, the model uses Linear Discriminant Analysis (LDA).
If \code{lambda > 0}, the model uses Penalized Discriminant Analysis (PDA).}
}
\value{
A \code{pp_strategy} object.
}
\description{
Creates a Penalized Discriminant Analysis (PDA) projection pursuit strategy
for use with \code{\link{pptr}} or \code{\link{pprf}}.
}
\examples{
# PDA with lambda = 0.5
pp_pda(0.5)

# Use with pptr
pptr(Type ~ ., data = iris, pp = pp_pda(0.5))

# Use with pprf
pprf(Type ~ ., data = iris, pp = pp_pda(0.5), vars = vars_uniform(n_vars = 2))

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}, \code{\link{vars_uniform}}, \code{\link{vars_all}}, \code{\link{cutpoint_mean_of_means}}
}
