% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{PPTree}
\alias{PPTree}
\title{Trains a Project-Pursuit oblique decision tree.}
\usage{
PPTree(formula = NULL, data = NULL, x = NULL, y = NULL)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x1 + x2 + ...}, where \code{y} is a vector of labels and \code{x1}, \code{x2}, ... are the features.}

\item{data}{A data frame containing the variables in the formula.}

\item{x}{A matrix containing the features for each observation.}

\item{y}{A matrix containing the labels for each observation.}
}
\value{
A PPTree model trained on \code{x} and \code{y}.
}
\description{
This function trains a Project-Pursuit oblique decision tree (PPTree) using either a formula and data frame interface or a matrix-based interface. When using the formula interface, specify the model formula and the data frame containing the variables. For the matrix-based interface, provide matrices for the features and labels directly.
}
\examples{
# Example 1: matrix interface with the `iris` dataset
PPTree(x = iris[, 1:4], y = iris[, 5])

# Example 2: formula interface with the `iris` dataset
PPTree(Species ~ ., data = iris)

# Example 3: matrix interface with the `crabs` dataset
x <- crabs[, c(2, 4:8)]
x$sex <- as.numeric(as.factor(x$sex))
PPTree(x = x, y = crabs[, 1])

# Example 4: formula interface with the `crabs` dataset
PPTree(sp ~ . - sex + as.numeric(as.factor(sex)), data = crabs)

}
