% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPForest.R
\name{print.PPForest}
\alias{print.PPForest}
\title{Prints a PPForest model.}
\usage{
\method{print}{PPForest}(x, ...)
}
\arguments{
\item{x}{A PPForest model.}

\item{...}{(unused) other parameters tipically passed to print}
}
\description{
Prints a PPForest model.
}
\examples{
model <- PPForest(Species ~ ., data = iris)
print(model)

}
