% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprf.R
\name{print.pprf}
\alias{print.pprf}
\title{Prints a compact summary of a pprf forest.}
\usage{
\method{print}{pprf}(x, ...)
}
\arguments{
\item{x}{A \code{pprf} model.}

\item{...}{Unused.}
}
\description{
Prints a compact summary of a pprf forest.
}
\seealso{
\code{\link{summary.pprf}}
}
