% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptr.R
\name{print.pptr}
\alias{print.pptr}
\title{Prints the structure of a pptr tree.}
\usage{
\method{print}{pptr}(x, ...)
}
\arguments{
\item{x}{A \code{pptr} model.}

\item{...}{Unused.}
}
\description{
Prints the structure of a pptr tree.
}
