% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{print.PPTree}
\alias{print.PPTree}
\title{Prints a PPTree model.}
\usage{
\method{print}{PPTree}(x, ...)
}
\arguments{
\item{x}{A PPTree model.}

\item{...}{(unused) other parameters typically passed to print.}
}
\description{
Prints a PPTree model.
}
\examples{
model <- PPTree(Type ~ ., data = iris)
print(model)

}
\seealso{
\code{\link{PPTree}}, \code{\link{summary.PPTree}}
}
