% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{sr_mean_of_means}
\alias{sr_mean_of_means}
\title{Mean-of-means split rule strategy.}
\usage{
sr_mean_of_means()
}
\value{
A \code{sr_strategy} object.
}
\description{
Creates a split rule strategy that splits at the midpoint between group
means. This is the default (and currently only) split rule.
}
\examples{
sr_mean_of_means()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}, \code{\link{pp_pda}}, \code{\link{dr_uniform}}
}
