% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-strategy.R
\name{stop_min_size}
\alias{stop_min_size}
\title{Minimum-size stopping rule.}
\usage{
stop_min_size(min_size = 5L)
}
\arguments{
\item{min_size}{Minimum node size to allow a split (default: 5).}
}
\value{
A \code{stop_strategy} object.
}
\description{
Creates a stopping rule that stops splitting when a node has fewer than
\code{min_size} observations. Used primarily for regression trees.
}
\examples{
stop_min_size(5)

}
\seealso{
\code{\link{stop_min_variance}}, \code{\link{stop_any}}
}
