% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop-strategy.R
\name{stop_pure_node}
\alias{stop_pure_node}
\title{Pure-node stopping rule.}
\usage{
stop_pure_node()
}
\value{
A \code{stop_strategy} object.
}
\description{
Creates a stopping rule that stops splitting when a node contains only one
group. This is the default stopping rule.
}
\examples{
stop_pure_node()

}
\seealso{
\code{\link{pptr}}, \code{\link{pprf}}
}
