% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPForest.R
\name{summary.PPForest}
\alias{summary.PPForest}
\title{Summarizes a PPForest model.}
\usage{
\method{summary}{PPForest}(object, ...)
}
\arguments{
\item{object}{A PPForest model.}

\item{...}{(unused) other parameters typically passed to summary.}
}
\description{
Summarizes a PPForest model.
}
\examples{
model <- PPForest(Type ~ ., data = iris)
summary(model)

}
\seealso{
\code{\link{PPForest}}, \code{\link{predict.PPForest}}, \code{\link{print.PPForest}}
}
