% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprf.R
\name{summary.pprf}
\alias{summary.pprf}
\title{Summarizes a pprf model.}
\usage{
\method{summary}{pprf}(object, ...)
}
\arguments{
\item{object}{A pprf model.}

\item{...}{(unused) other parameters typically passed to summary.}
}
\description{
Summarizes a pprf model.
}
\examples{
model <- pprf(Type ~ ., data = iris)
summary(model)

}
\seealso{
\code{\link{pprf}}, \code{\link{predict.pprf}}, \code{\link{print.pprf}}
}
