% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptr.R
\name{summary.pptr}
\alias{summary.pptr}
\title{Summarizes a pptr model.}
\usage{
\method{summary}{pptr}(object, ...)
}
\arguments{
\item{object}{A pptr model.}

\item{...}{(unused) other parameters typically passed to summary.}
}
\description{
Summarizes a pptr model.
}
\examples{
model <- pptr(Type ~ ., data = iris)
summary(model)

}
\seealso{
\code{\link{pptr}}, \code{\link{predict.pptr}}, \code{\link{print.pptr}}
}
