% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars-strategy.R
\name{vars_all}
\alias{vars_all}
\title{All-variables selection strategy.}
\usage{
vars_all()
}
\value{
A \code{vars_strategy} object.
}
\description{
Creates a variable selection strategy that uses all variables at each
split. This is the default for single trees (\code{\link{pptr}}).
}
\examples{
vars_all()

}
\seealso{
\code{\link{pptr}}, \code{\link{vars_uniform}}, \code{\link{pp_pda}}, \code{\link{cutpoint_mean_of_means}}
}
