% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{load_json}
\alias{load_json}
\title{Load a model from a JSON file.}
\usage{
load_json(path)
}
\arguments{
\item{path}{File path to read the JSON from.}
}
\value{
A \code{pptr} or \code{pprf} model.
}
\description{
Deserializes a \code{pptr} or \code{pprf} model from a JSON file. The model
can be used for prediction immediately. If variable importance metrics were
saved, they are restored as well.
}
\details{
Note that \code{formula}, \code{x}, and \code{y} are not stored in the JSON
and will be \code{NULL} on the loaded model. This means formula-based
\code{predict()} (passing a data frame) will not work; pass a numeric matrix
instead.
}
\examples{
model <- pptr(Type ~ ., data = iris, seed = 0)
path <- tempfile(fileext = ".json")
save_json(model, path)
loaded <- load_json(path)
predict(loaded, as.matrix(iris[, 1:4]))

}
\seealso{
\code{\link{save_json}}, \code{\link{pptr}}, \code{\link{pprf}}
}
