% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\name{pp_tree}
\alias{pp_tree}
\title{Parsnip model specification for PPTree.}
\usage{
pp_tree(mode = "classification", penalty = NULL)
}
\arguments{
\item{mode}{A character string for the model type. Only \code{"classification"} is supported.}

\item{penalty}{The regularization parameter (maps to \code{lambda}).}
}
\value{
A parsnip model specification.
}
\description{
Creates a model specification for a single Projection Pursuit decision tree.
Use \code{set_engine("PPTree")} to select the PPTree engine.
}
