% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{PPTree}
\alias{PPTree}
\title{Trains a Project-Pursuit oblique decision tree.}
\usage{
PPTree(x, y)
}
\arguments{
\item{x}{A matrix containing the features for each observation.}

\item{y}{A matrix containing the labels for each observation.}
}
\value{
A PPTree model trained on \code{x} and \code{y}.
}
\description{
Trains a Project-Pursuit oblique decision tree.
}
\examples{
# Example 1: with the preprocessed `iris` dataset
library(PPTree)
x <- as.matrix(PPTreeIris[, 1:4])
y <- as.matrix(PPTreeIris[, 5])
PPTree(x, y)

# Example 2: with the preprocessed `crabs` dataset
library(PPTree)
x <- as.matrix(PPTreeCrabs[, 2:7])
y <- as.matrix(PPTreeCrabs[, 1])
PPTree(x, y)

}
