% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{PPTree}
\alias{PPTree}
\title{Trains a Project-Pursuit oblique decision tree.}
\usage{
PPTree(x, y)
}
\arguments{
\item{x}{A matrix containing the features for each observation.}

\item{y}{A matrix containing the labels for each observation.}
}
\value{
A PPTree model trained on \code{x} and \code{y}.
}
\description{
Trains a Project-Pursuit oblique decision tree.
}
\examples{
# Example 1: with the `iris` dataset
PPTree(x = iris[, 1:4], y = iris[, 5])

# Example 2: with the `crabs` dataset
x <- crabs[, c(2, 4:8)]
x$sex <- as.numeric(as.factor(x$sex))
PPTree(x = x, y = crabs[, 1])

}
