% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{PPTree}
\alias{PPTree}
\title{Trains a Project-Pursuit oblique decision tree.}
\usage{
PPTree(formula = NULL, data = NULL, x = NULL, y = NULL, lambda = 0)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x1 + x2 + ...}, where \code{y} is a vector of labels and \code{x1}, \code{x2}, ... are the features.}

\item{data}{A data frame containing the variables in the formula.}

\item{x}{A matrix containing the features for each observation.}

\item{y}{A matrix containing the labels for each observation.}

\item{lambda}{A regularization parameter. If \code{lambda = 0}, the model is trained using Linear Discriminant Analysis (LDA). If \code{lambda > 0}, the model is trained using Penalized Discriminant Analysis (PDA).}
}
\value{
A PPTree model trained on \code{x} and \code{y}.
}
\description{
This function trains a Project-Pursuit oblique decision tree (PPTree) using either a formula and data frame interface or a matrix-based interface. When using the formula interface, specify the model formula and the data frame containing the variables. For the matrix-based interface, provide matrices for the features and labels directly.
If \code{lambda = 0}, the model is trained using Linear Discriminant Analysis (LDA). If \code{lambda > 0}, the model is trained using Penalized Discriminant Analysis (PDA).
}
\examples{

# Example 1: formula interface with the `iris` dataset
PPTree(Type ~ ., data = iris)

# Example 2: formula interface with the `iris` dataset with regularization
PPTree(Type ~ ., data = iris, lambda = 0.5)

# Example 3: matrix interface with the `iris` dataset
PPTree(x = iris[, 1:4], y = iris[, 5])

# Example 4: matrix interface with the `iris` dataset with regularization
PPTree(x = iris[, 1:4], y = iris[, 5], lambda = 0.5)

# Example 5: formula interface with the `crabs` dataset
PPTree(Type ~ . - sex + as.numeric(as.factor(sex)), data = crabs)

# Example 6: formula interface with the `crabs` dataset with regularization
PPTree(Type ~ . - sex + as.numeric(as.factor(sex)), data = crabs, lambda = 0.5)

# Example 7: matrix interface with the `crabs` dataset
x <- crabs[, 2:5]
x$sex <- as.numeric(as.factor(crabs$sex))
PPTree(x = x, y = crabs$Type)

# Example 8: matrix interface with the `crabs` dataset with regulartion
x <- crabs[, 2:5]
x$sex <- as.numeric(as.factor(crabs$sex))
PPTree(x = x, y = crabs$Type, lambda = 0.5)

}
