% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{predict.PPTree}
\alias{predict.PPTree}
\title{Predicts the labels of a set of observations using a PPTree model.}
\usage{
\method{predict}{PPTree}(model, x)
}
\arguments{
\item{model}{A PPTree model.}

\item{x}{A matrix containing the features for each new observation.}
}
\value{
A matrix containing the predicted labels for each observation.
}
\description{
Predicts the labels of a set of observations using a PPTree model.
}
\examples{
# Example 1: with the preprocessed `iris` dataset
library(PPTree)
x <- as.matrix(PPTreeIris[, 1:4])
y <- as.matrix(PPTreeIris[, 5])
model <- PPTree(x, y)
predict(model, x)

# Example 2: with the preprocessed `crabs` dataset
library(PPTree)
x <- as.matrix(PPTreeCrabs[, 2:7])
y <- as.matrix(PPTreeCrabs[, 1])
model <- PPTree(x, y)
predict(model, x)

}
