% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{predict.PPTree}
\alias{predict.PPTree}
\title{Predicts the labels of a set of observations using a PPTree model.}
\usage{
\method{predict}{PPTree}(model, x)
}
\arguments{
\item{model}{A PPTree model.}

\item{x}{A matrix containing the features for each new observation.}
}
\value{
A matrix containing the predicted labels for each observation.
}
\description{
Predicts the labels of a set of observations using a PPTree model.
}
\examples{
# Example 1: with the `iris` dataset
model <- PPTree(iris[, 1:4], iris[, 5])
predict(model, iris[, 1:4])

# Example 2: with the `crabs` dataset
x <- crabs[, c(2, 4:8)]
x$sex <- as.numeric(as.factor(x$sex))
model <- PPTree(x = x, y = crabs[, 1])
predict(model, x)

}
