% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{predict.PPTree}
\alias{predict.PPTree}
\title{Predicts the labels of a set of observations using a PPTree model.}
\usage{
\method{predict}{PPTree}(model, x)
}
\arguments{
\item{model}{A PPTree model.}

\item{x}{A matrix containing the features for each new observation.}
}
\value{
A matrix containing the predicted labels for each observation.
}
\description{
Predicts the labels of a set of observations using a PPTree model.
}
\examples{
# Example 1: with the `iris` dataset
model <- PPTree(Species ~ ., data = iris)
predict(model, iris)

# Example 2: with the `crabs` dataset
model <- PPTree(sp ~ . - sex + as.numeric(as.factor(sex)), data = crabs)
predict(model, crabs)

}
