% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{predict.PPTree}
\alias{predict.PPTree}
\title{Predicts the labels of a set of observations using a PPTree model.}
\usage{
\method{predict}{PPTree}(object, ...)
}
\arguments{
\item{object}{A PPTree model.}

\item{...}{(unused) other parameters tipically passed to predict.}
}
\value{
A matrix containing the predicted labels for each observation.
}
\description{
Predicts the labels of a set of observations using a PPTree model.
}
\examples{
# Example 1: with the `iris` dataset
model <- PPTree(Type ~ ., data = iris)
predict(model, iris)

# Example 2: with the `crabs` dataset
model <- PPTree(Type ~ . - sex + as.numeric(as.factor(sex)), data = crabs)
predict(model, crabs)

}
