% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprf.R
\name{print.pprf}
\alias{print.pprf}
\title{Prints a pprf model.}
\usage{
\method{print}{pprf}(x, ...)
}
\arguments{
\item{x}{A pprf model.}

\item{...}{(unused) other parameters typically passed to print.}
}
\description{
Prints a pprf model.
}
\examples{
model <- pprf(Type ~ ., data = iris)
print(model)

}
\seealso{
\code{\link{pprf}}, \code{\link{summary.pprf}}
}
