% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptr.R
\name{print.pptr}
\alias{print.pptr}
\title{Prints a pptr model.}
\usage{
\method{print}{pptr}(x, ...)
}
\arguments{
\item{x}{A pptr model.}

\item{...}{(unused) other parameters typically passed to print.}
}
\description{
Prints a pptr model.
}
\examples{
model <- pptr(Type ~ ., data = iris)
print(model)

}
\seealso{
\code{\link{pptr}}, \code{\link{summary.pptr}}
}
