% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{print.PPTree}
\alias{print.PPTree}
\title{Prints a PPTree model.}
\usage{
\method{print}{PPTree}(model)
}
\arguments{
\item{model}{A PPTree model.}
}
\description{
Prints a PPTree model.
}
\examples{
model <- PPTree(Species ~ ., data = iris)
print(model)

}
