% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPForest.R
\name{summary.PPForest}
\alias{summary.PPForest}
\title{Summarizes a PPForest model.}
\usage{
\method{summary}{PPForest}(object, ...)
}
\arguments{
\item{object}{A PPForest model.}

\item{...}{(unused) other parameters tipically passed to print}
}
\description{
Summarizes a PPForest model.
}
\examples{
model <- PPForest(Species ~ ., data = iris)
summary(model)

}
