% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprf.R
\name{summary.pprf}
\alias{summary.pprf}
\title{Summary of a pprf forest (shared header + VI).}
\usage{
\method{summary}{pprf}(object, ...)
}
\arguments{
\item{object}{A \code{pprf} model.}

\item{...}{Unused.}
}
\description{
Summary of a pprf forest (shared header + VI).
}
