% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPTree.R
\name{summary.PPTree}
\alias{summary.PPTree}
\title{Summarizes a PPTree model.}
\usage{
\method{summary}{PPTree}(object, ...)
}
\arguments{
\item{object}{A PPTree model.}

\item{...}{(unused) other parameters tipically passed to print}
}
\description{
Summarizes a PPTree model.
}
\examples{
model <- PPTree(Type ~ ., data = iris)
summary(model)

}
