% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Dataset}
\format{
A data frame with 178 rows and 14 variables
}
\usage{
data(wine)
}
\description{
A dataset containing 178 observations from 3 wine grown cultivares in Italy.
}
\details{
\describe{
  \item{Type}{ Class variable has 3 classes that are 3 different wine grown cultivares in Italy. }
  \item{X1 to X13}{Check vbles}
}
}
\keyword{datasets}
